feat: UIG-2939 - integration lib - componenten
Integratie componenten verplaatst naar integration lib
Storybook en Cypress
