+++ apps/storybook-e2e/src/e2e/ontwerp/form/demo/form-demo.cy.ts
@@ -1,0 +1,9 @@
+const formDemoUrl = 'http://localhost:8080/iframe.html?id=ontwerp-form-demo--demo&viewMode=story';
+
+describe('story - form demo', () => {
+    it('should render', () => {
+        cy.visit(formDemoUrl);
+
+        cy.get('vl-form-demo').shadow();
+    });
+});
