--- apps/storybook-e2e/src/e2e/ontwerp/popover-menu/popover-menu.cy.ts
@@ -1,27 +1,0 @@
-const PopoverMenuHamburgerUrl =
-    'http://localhost:8080/iframe.html?args=&id=ontwerp-popover-menu--hamburger-menu&viewMode=story';
-const PopoverMenuKebabrUrl =
-    'http://localhost:8080/iframe.html?args=&id=ontwerp-popover-menu--kebab-menu&viewMode=story';
-const PopoverMenuMeatballUrl =
-    'http://localhost:8080/iframe.html?args=&id=ontwerp-popover-menu--meatball-menu&viewMode=story';
-
-describe('story popover-hamburger-menu', () => {
-    it('should display story', () => {
-        cy.visit(PopoverMenuHamburgerUrl);
-        cy.get('vl-popover').shadow();
-    });
-});
-
-describe('story popover-kebab-menu', () => {
-    it('should display story', () => {
-        cy.visit(PopoverMenuKebabrUrl);
-        cy.get('vl-popover').shadow();
-    });
-});
-
-describe('story popover-meatball-menu', () => {
-    it('should display story', () => {
-        cy.visit(PopoverMenuMeatballUrl);
-        cy.get('vl-popover').shadow();
-    });
-});
