+++ apps/storybook-e2e/src/e2e/ontwerp/popover/menu/popover-menu.cy.ts
@@ -1,0 +1,9 @@
+const popoverMenuUrl = 'http://localhost:8080/iframe.html?id=ontwerp-popover-menu--menu&viewMode=story';
+
+describe('story - popover menu', () => {
+    it('should render', () => {
+        cy.visit(popoverMenuUrl);
+
+        cy.get('vl-popover-menu').find('vl-popover').shadow();
+    });
+});
