--- apps/storybook/docs/e_ontwerp/form/1_form-documentatie/form.stories.mdx
@@ -1,11 +1,0 @@
-import { Meta } from '@storybook/addon-docs';
-
-<Meta title="Ontwerp/Form"/>
-
-# Form
-
-## Inhoudstafel
-
-- [Form Demo](/docs/ontwerp-form-demo--documentatie)
-- [FormData](/docs/ontwerp-form-formdata--documentatie)
-- [Custom Validatie](/docs/ontwerp-form-custom-validation--documentatie)
