--- apps/storybook/docs/e_ontwerp/form/2_form-demo/form-demo.stories-doc.mdx
@@ -1,54 +1,0 @@
-import { Meta } from '@storybook/addon-docs';
-import { Canvas, Source } from '@storybook/blocks';
-import * as formDemoStories from './form-demo.stories';
-import formDemoSource from '!!raw-loader!./form-demo.component';
-
-<Meta of={formDemoStories} />
-
-# Form Demo
-
-Dit is een voorbeeld van hoe de form componenten in een form gebruikt kunnen worden.<br/>
-De submitted waarden worden in deze demo in de console gelogd.
-
-## Componenten
-
-- [vl-input-field-next](/docs/form-next-input-field--documentatie)
-- [vl-input-field-masked-next](/docs/form-next-input-field-masked--documentatie)
-- [vl-textarea-next](/docs/form-next-textarea--documentatie)
-- [vl-datepicker-next](/docs/form-next-datepicker--documentatie)
-- [vl-select-next](/docs/form-next-select--documentatie)
-- [vl-radio-group-next](/docs/form-next-radio-group--documentatie)
-- [vl-upload-next](/docs/form-next-upload--documentatie)
-- [vl-checkbox-next](/docs/form-next-checkbox--documentatie)
-- [vl-error-message-next](/docs/form-next-error-message--documentatie)
-- [vl-form-label-next](/docs/form-next-form-label--documentatie)
-
-## Native elementen
-
-Gebruik voor de volgende elementen native HtmlElements met css classes:
-- Form: native HTMLFormElement met class `vl-form`
-- Button: native HtmlButtonElement met class `vl-button`
-- Action Group: native HtmlDivElement met class `vl-action-group`
-
-## Grid
-
-Maak voor het opbouwen van de grid structuur gebruik van de [vl-form-grid classes](https://overheid.vlaanderen.be/webuniversum/v3/documentation/form-grid) van Digitaal Vlaanderen.<br/>
-Plaats deze op native HtmlDivElements.
-
-## FormData
-
-Gebruik de [FormData](https://developer.mozilla.org/en-US/docs/Web/API/FormData) API om de form data te verzamelen.
-Voor velden die meerdere waardes per naam kunnen hebben, zoals bv. de `vl-upload-next`, kan je de `FormData.getAll()` methode gebruiken.
-
-## ❗Let op
-
-Indien je een form control wrapped in een custom component, zorg er voor dat dit custom component geen Shadow DOM heeft.<br/>
-Als dit custom component een Shadow DOM heeft, zal het native HtmlFormElement de form control niet meer vinden.
-
-<Canvas of={formDemoStories.Demo} sourceState='none' />
-
-<details open>
-    <summary>Code</summary>
-    <Source code={formDemoSource} language="ts" dark={true} />
-</details>
-
