--- apps/storybook/docs/e_ontwerp/form/3_form-data/form-data.component.uig-css.ts
@@ -1,10 +1,0 @@
-import { css, CSSResult } from 'lit';
-
-const styles: CSSResult = css`
-    section.form-data {
-        display: flex;
-        flex-direction: column;
-        gap: 1rem;
-    }
-`;
-export default styles;
