+++ apps/storybook/docs/e_ontwerp/form/3_form-data/form-data.stories-doc.mdx
@@ -4,34 +4,34 @@ import * as formDataStories from './form-data.stories';
 
 <Meta of={formDataStories} />
 
-# FormData
+# Form - Data
 
-Om de FormData te verzamelen, kan je de [FormData](https://developer.mozilla.org/en-US/docs/Web/API/FormData) API gebruiken.
+Om de form data te verzamelen, kan je de [FormData](https://developer.mozilla.org/en-US/docs/Web/API/FormData) API gebruiken.
 
-Echter, deze API is niet altijd even handig in gebruik. Als je `FormData.entries()` gebruikt, dan krijg je een iterator van de waardes. Dit is voldoende voor enkelvoudige waardes, maar voor inputs met meervoudige waardes krijg je enkel de laatst gekozen waarde.
+Echter, deze API is niet altijd even handig in gebruik. Als je `FormData.entries()` gebruikt, dan krijg je een iterator van de waardes.
+Dit is voldoende voor inputs met 1 waarde, maar voor inputs met meerdere waardes krijg je enkel de laatst gekozen waarde terug.
 
 Daarom leek het ons aangewezen dat je altijd met een consistent object kan werken:
-- voor inputs met 1 waarde (zoals een input), krijg je altijd een enkele waarde
-- voor inputs met meerdere waardes (zoals een multiselect), krijg je altijd een array van waardes
+- voor inputs met 1 waarde, krijg je altijd een enkele waarde
+- voor inputs met meerdere waardes (bv. multiselect), krijg je altijd een array van waardes
 
 ## parseFormData(formElement: Form, multiFormControlNames?: string[])
 
-We hebben een helper functie voorzien om de FormData te parsen naar een object met de waardes van de velden.
+We hebben een helper functie voorzien om de form data te parsen naar een object met de waardes van de velden.
 
 Deze functie aanvaard 2 parameters:
-- `formElement: Form`: het Form element waarvan je de FormData wil parsen
-- `multiFormControlNames: string[]`: een array van form control namen die meerdere waarden kunnen hebben. Optioneel, als deze parameter niet wordt meegegeven, dan wordt er gezocht naar alle form controls die meerdere waarden kunnen hebben.
+- `formElement: Form`: het Form element waarvan je de data wil parsen
+- `multiFormControlNames?: string[]`: een array van form control namen die meerdere waardes kunnen hebben. Als deze parameter niet wordt meegegeven, dan wordt er gezocht naar alle form controls die meerdere waardes kunnen hebben.
 
-### Gebruik:
+### Gebruik
 
 ```typescript
 import { parseFormData } from "@domg-wc/form/utils";
 
 const form = document.querySelector('form');
-console.log(parseFormData(form));
+const data = parseFormData(form);
 ```
 
 ### Voorbeeld
 
-<Canvas of={formDataStories.FormDataStory} sourceState='none' />
-
+<Canvas of={formDataStories.FormData} sourceState='none' />
