+++ apps/storybook/docs/e_ontwerp/form/3_form-data/form-data.stories.ts
@@ -2,14 +2,13 @@ import { html } from 'lit-html';
 import { Meta } from '@storybook/web-components';
 import formDataDoc from './form-data.stories-doc.mdx';
 import { registerWebComponents } from '@domg-wc/common-utilities';
-import { FormDataComponent } from './form-data.component';
+import { VlFormDataComponent } from '@domg-wc/integration/form/form-data';
 
-registerWebComponents([FormDataComponent]);
+registerWebComponents([VlFormDataComponent]);
 
 export default {
-    title: 'Ontwerp/Form/FormData',
+    title: 'Ontwerp/Form/Form Data',
     tags: ['autodocs'],
-    component: 'form data',
     parameters: {
         docs: {
             page: formDataDoc,
@@ -17,4 +16,4 @@ export default {
     },
 } as Meta;
 
-export const FormDataStory = () => html`<vl-form-data></vl-form-data>`;
+export const FormData = () => html`<vl-form-data></vl-form-data>`;
