+++ apps/storybook/docs/e_ontwerp/form/4_custom-validation/form-custom-validation.stories-doc.mdx
@@ -1,0 +1,27 @@
+import { Meta } from '@storybook/addon-docs';
+import { Canvas, Source } from '@storybook/blocks';
+import * as customValidationStories from './form-custom-validation.stories';
+import customValidationSource from '!!raw-loader!@domg-wc/integration/form/custom-validation/vl-form-custom-validation.component';
+
+<Meta of={customValidationStories} />
+
+# Form - Custom Validatie
+
+Om gebruik te maken van custom validatie moet je een nieuwe component maken dat overerft van één van onze form controls.<br/>
+Aan dit component kan je je custom validator toevoegen. Via de `customError` ValidityState key kan je een error message linken aan je custom validator.
+
+## Extra informatie
+
+Het is nodig om een custom component te maken voor elke custom validator die je wil toevoegen:<br/>
+- De validators worden toegevoegd op de form controls d.m.v. een statische array. Dit betekent dat als je een custom validator toevoegt aan een form control, deze validator zal toegevoegd worden aan alle form controls van dit type.<br/>
+- De form controls maken gebruik van de ValidityState API waardoor er maximum 1 custom validator toegevoegd kan worden aan een form control.<br/>
+Bij het toevoegen van een 2de custom validator zal de eerste custom validator overschreven worden.<br/>
+
+## Voorbeeld
+
+<Canvas of={customValidationStories.CustomValidation} sourceState='none' />
+
+<details open>
+    <summary>Code</summary>
+    <Source code={customValidationSource} language="ts" dark={true} />
+</details>
