+++ apps/storybook/docs/e_ontwerp/popover/1_menu/popover-menu.stories-doc.mdx
@@ -1,0 +1,34 @@
+import { Meta } from '@storybook/addon-docs';
+import { Canvas, Source } from '@storybook/blocks';
+import * as menuStories from './popover-menu.stories'
+import menuSource from '!!raw-loader!@domg-wc/integration/popover/menu/vl-popover-menu.component';
+
+<Meta of={menuStories} />
+
+# Popover - Menu
+
+Een popover-menu wordt gebruikt om extra acties weer te geven die niet direct zichtbaar zijn.<br/>
+Het wordt meestal afgebeeld mbv een kebab menu icoon (drie verticale punten).
+
+## Gebruik
+
+Het popover-menu wordt gebruikt wanneer er beperkte ruimte is en niet alle acties direct weergeven kunnen worden.
+
+## Ontwerp
+
+- Icoon: Het menu wordt meestal afgebeeld mbv een kebab menu icoon (drie verticale punten).
+
+- Plaatsing: Het menu kan worden geplaatst in de linkerbovenhoek, rechterbovenhoek of elders op de pagina, afhankelijk
+  van het ontwerp en de context.
+
+- Interactie: Door op het icoon te klikken, wordt het menu geopend en de extra acties weergegeven.
+  Het menu kan gesloten worden door opnieuw op het icoon te klikken, of ergens buiten het menu.
+
+## Voorbeeld
+
+<Canvas of={menuStories.Menu} sourceState='none' />
+
+<details open>
+    <summary>Code</summary>
+    <Source code={menuSource} language="ts" dark={true} />
+</details>
