+++ apps/storybook/docs/e_ontwerp/popover/2_menu-accordion/popover-menu-accordion.stories-doc.mdx
@@ -1,0 +1,34 @@
+import { Meta } from '@storybook/addon-docs';
+import { Canvas, Source } from '@storybook/blocks';
+import * as menuAccordionStories from './popover-menu-accordion.stories'
+import menuAccordionSource from '!!raw-loader!@domg-wc/integration/popover/menu-accordion/vl-popover-menu-accordion.component';
+
+<Meta of={menuAccordionStories} />
+
+# Popover - Menu Accordion
+
+De accordion wordt gebruikt om inhoud te organiseren in een uitklapbaar formaat.<br/>
+Elke accordion kan een popover-menu bevatten dat specifieke acties aanbiedt.
+
+## Gebruik
+
+De accordion met popover-menu wordt gebruikt om inhoud te structureren en gebruikers de mogelijkheid te geven om specifieke acties uit te voeren.
+
+## Ontwerp
+
+- Accordion: elke accordion bestaat uit een titel en een inhoudsgebied dat kan worden uit- of ingeklapt.
+
+- Popover-menu: elke accordion kan een popover-menu bevatten dat specifieke acties aanbiedt.
+  Het menu wordt meestal afgebeeld mbv een kebab menu icoon (drie verticale punten).
+
+- Interactie: gebruikers kunnen op de titel klikken om de accordion uit- of in te klappen.
+  Door op het popover-menu icoon te klikken wordt het menu geopend en de extra acties weergegeven.
+
+## Voorbeeld
+
+<Canvas of={menuAccordionStories.MenuAccordion} sourceState='none'/>
+
+<details open>
+    <summary>Code</summary>
+    <Source code={menuAccordionSource} language="ts" dark={true} />
+</details>
