+++ cypress.config.ts
@@ -26,6 +26,7 @@ export default defineConfig({
                         '@domg-wc/common-utilities': path.resolve(__dirname, 'libs/common/utilities/src/index.ts'),
                         '@domg-wc/components': path.resolve(__dirname, 'libs/components/src/index.ts'),
                         '@domg-wc/elements': path.resolve(__dirname, 'libs/elements/src/index.ts'),
+                        '@domg-wc/form': path.resolve(__dirname, 'libs/form/src/'),
                         '@domg-wc/map': path.resolve(__dirname, 'libs/map/src/index.ts'),
                         '@domg-wc/sections': path.resolve(__dirname, 'libs/sections/src/index.ts'),
                         '@domg-wc/qlik': path.resolve(__dirname, 'libs/qlik/src/index.ts'),
