+++ libs/integration/project.json
@@ -14,9 +14,12 @@
                 "assets": ["libs/integration/*.md"]
             }
         },
-        "publish": {
-            "command": "node tools/scripts/publish.mjs integration {args.ver} {args.tag}",
-            "dependsOn": ["build"]
+        "build-with-js": {
+            "executor": "nx:run-commands",
+            "dependsOn": ["build"],
+            "options": {
+                "command": "node tools/build-with-js/copy-integration-js.mjs"
+            }
         },
         "lint": {
             "executor": "@nx/eslint:lint"
@@ -25,7 +28,8 @@
             "executor": "@nx/jest:jest",
             "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
             "options": {
-                "jestConfig": "libs/integration/jest.config.ts"
+                "jestConfig": "libs/integration/jest.config.ts",
+                "passWithNoTests": true
             }
         }
     },
