+++ libs/integration/src/form/custom-validation/index.ts
@@ -1,0 +1,1 @@
+export { VlFormCustomValidationComponent } from './vl-form-custom-validation.component';
