+++ libs/integration/src/form/demo/index.ts
@@ -1,0 +1,1 @@
+export { VlFormDemoComponent } from './vl-form-demo.component';
