+++ libs/integration/src/form/form-data/index.ts
@@ -1,0 +1,1 @@
+export { VlFormDataComponent } from './vl-form-data.component';
