--- libs/integration/src/lib/integration.spec.ts
@@ -1,7 +1,0 @@
-import { integration } from './integration';
-
-describe('integration', () => {
-    it('should work', () => {
-        expect(integration()).toEqual('integration');
-    });
-});
