--- libs/integration/src/lib/integration.ts
@@ -1,3 +1,0 @@
-export function integration(): string {
-    return 'integration';
-}
