+++ libs/integration/src/popover/menu-accordion/index.ts
@@ -1,0 +1,1 @@
+export { VlPopoverMenuAccordionComponent } from './vl-popover-menu-accordion.component';
