+++ libs/integration/src/popover/menu/index.ts
@@ -1,0 +1,1 @@
+export { VlPopoverMenuComponent } from './vl-popover-menu.component';
