+++ libs/integration/tsconfig.lib.json
@@ -3,8 +3,9 @@
     "compilerOptions": {
         "outDir": "../../dist/out-tsc",
         "declaration": true,
-        "types": ["node"]
+        "types": ["cypress", "cypress-axe"]
     },
+    "files": ["../../cypress/types.d.ts"],
     "include": ["src/**/*.ts"],
     "exclude": ["jest.config.ts", "src/**/*.spec.ts", "src/**/*.test.ts"]
 }
